# docs/swagger.yaml
basePath: /
definitions:
  controllers.AccessRequest:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
      expires_at:
        type: string
      user_id:
        type: string
    required:
    - entity_id
    - entity_type
    - expires_at
    - user_id
    type: object
  controllers.AddAreaRequest:
    properties:
      area:
        type: string
    required:
    - area
    type: object
...
  controllers.RemoveAreaRequest:
    properties:
      area:
        type: string
    required:
    - area
    type: object
...
  controllers.UpdateAreaRequest:
    properties:
      new_area:
        type: string
      old_area:
        type: string
    required:
    - new_area
    - old_area
    type: object
...
  models.Caso:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      direcciones:
        items:
          $ref: '#/definitions/models.Direccion'
        type: array
      documento:
        items:
          $ref: '#/definitions/models.Documento'
        type: array
      empresas:
        items:
          $ref: '#/definitions/models.Empresa'
        type: array
      id:
        type: string
      iio:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
      modalidad:
        type: string
      nombre:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tie:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      vanalista:
        type: integer
      vcoordinador:
        type: integer
      vdirector:
        type: integer
      vehiculos:
        items:
          $ref: '#/definitions/models.Vehiculo'
        type: array
    type: object
  models.Correo:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Direccion:
    properties:
      area:
        type: string
      created_at:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      estado:
        type: string
      id:
        type: string
      lugar:
        type: string
      municipio:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
    type: object
  models.Documento:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      documento:
        type: string
      id:
        type: string
      nombre:
        type: string
      numero:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Empresa:
    properties:
      actividad_economica:
        type: string
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      id:
        type: string
      nombre:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      rif:
        type: string
      socios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.IIO:
    properties:
      area:
        type: string
      created_at:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      imagen_url:
        type: string
      lugar:
        type: string
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      redi:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tie:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      verificada:
        type: boolean
      zodi:
        type: string
    type: object
...
  models.Nacionalidad:
    properties:
      area:
        type: string
      created_at:
        type: string
      gentilicio:
        type: string
      id:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Pasaporte:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      foto:
        type: string
      id:
        type: string
      numero:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Persona:
    properties:
      alias:
        type: string
      apellido:
        type: string
      area:
        type: string
      cargo_actual:
        type: string
      cedula:
        type: string
      correo:
        type: string
      correos:
        items:
          $ref: '#/definitions/models.Correo'
        type: array
      created_at:
        type: string
      direcciones:
        items:
          $ref: '#/definitions/models.Direccion'
        type: array
      empresas:
        items:
          $ref: '#/definitions/models.Empresa'
        type: array
      estado_civil:
        type: string
      filiacion_politica:
        type: string
      id:
        type: string
      ideologia:
        type: string
      iio:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
      informacion_de_interes:
        type: string
      nacionalidad:
        items:
          $ref: '#/definitions/models.Nacionalidad'
        type: array
      nombre:
        type: string
      pasaporte:
        items:
          $ref: '#/definitions/models.Pasaporte'
        type: array
      profesion:
        type: string
      redes:
        items:
          $ref: '#/definitions/models.Redes'
        type: array
      relacionados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      religion:
        type: string
      telefono:
        type: string
      tipo_perfil:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoraciones:
        type: string
      vehiculos:
        items:
          $ref: '#/definitions/models.Vehiculo'
        type: array
      visa:
        items:
          $ref: '#/definitions/models.Visa'
        type: array
    type: object
  models.Redes:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.TemporaryAccess:
    properties:
      created_at:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      expires_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      apellido:
        type: string
      area:
        type: string
      cedula:
        type: string
      correo:
        type: string
      created_at:
        type: string
      credencial:
        type: string
      hash:
        type: string
      id:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      otp_secret:
        type: string
      redi:
        type: string
      telefono:
        type: string
      tie:
        type: string
      updated_at:
        type: string
    type: object
  models.Vehiculo:
    properties:
      area:
        type: string
      año:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      linea:
        type: string
      marca:
        type: string
      matricula:
        type: string
      modelo:
        type: string
      numero:
        type: string
      propietario:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      vehiculo:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
    type: object
  models.Visa:
    properties:
      aprobada:
        type: boolean
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      id:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoracion:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API SISTEMA DE EXPLORACION POPULAR.
  title: API
  version: "1.0"
paths:
  /casos:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos del caso
        in: body
        name: caso
        required: true
        schema:
          $ref: '#/definitions/models.Caso'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Caso'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crea un nuevo caso
      tags:
      - Caso
  /casos/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del caso a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Caso eliminado exitosamente
        "404":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Elimina un caso por su ID
      tags:
      - Caso
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del caso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Caso'
        "404":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtiene un caso por su ID
      tags:
      - Caso
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del caso a actualizar
        in: path
        name: id
        required: true
        type: string
      - description: Datos del caso a actualizar
        in: body
        name: caso
        required: true
        schema:
          $ref: '#/definitions/models.Caso'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Caso'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualiza un caso existente por su ID
      tags:
      - Caso
  /casos/valorar/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del caso
        in: path
        name: id
        required: true
        type: string
      - description: Valoración
        in: body
        name: valor
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Caso'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Valorar un caso
      tags:
      - Caso
  /configuracion/acceso-temporal:
    post:
      consumes:
      - application/json
      description: Otorga acceso temporal a un usuario de otra área a una entidad
        específica
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos de acceso temporal
        in: body
        name: access
        required: true
        schema:
          $ref: '#/definitions/controllers.AccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TemporaryAccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Otorga acceso temporal a un usuario a una entidad específica
      tags:
      - configuracion
  /configuracion/area:
    delete:
      consumes:
      - application/json
      description: Elimina un área de la lista de áreas válidas
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Nombre del área
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/controllers.RemoveAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Elimina un área
      tags:
      - configuracion
    post:
      consumes:
      - application/json
      description: Agrega un área nueva a la lista de áreas válidas
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Nombre del área
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/controllers.AddAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Agrega un área nueva
      tags:
      - configuracion
    put:
      consumes:
      - application/json
      description: Actualiza el nombre de un área en la lista de áreas válidas
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos del área
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Actualiza un área
      tags:
      - configuracion
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"